+++ TodoBasicWithAuth/TodoApi.cs
@@ -26,7 +26,7 @@ namespace Todos
 
         public async Task GetAsync(TodoDbContext db, HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
@@ -52,7 +52,7 @@ namespace Todos
 
         public async Task DeleteAsync(TodoDbContext db, HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
