+++ TodoWithRoutes/TodoApi.cs
@@ -25,7 +25,7 @@ namespace Todos
 
         public async Task GetAsync(HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
@@ -53,7 +53,7 @@ namespace Todos
 
         public async Task DeleteAsync(HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
